modified posts controller

+++ controllers/posts.js
@@ -1,45 +1,40 @@
-/* const postsRouter = require('express').Router();
+const postsRouter = require('express').Router();
 const Post = require('../models/post');
 const User = require('../models/user');
 const jwt = require('jsonwebtoken');
 
 postsRouter.get('/', async (request, response) => {
   const posts = await Post.find({}).populate('user', { username: 1, name: 1 });
-  response.json(posts.map((post) => post.toJSON()));
-});
-
-postsRouter.get('/:id', async (request, response) => {
-  const post = await Post.findById(request.params.id);
-  if (post) {
-    response.json(post.toJSON());
-  } else {
-    response.status(404).end();
-  }
+  response.json(posts);
 });
 
 postsRouter.post('/', async (request, response) => {
-  const body = request.body;
+  const post = new Post(request.body);
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' });
   }
 
   const user = await User.findById(decodedToken.id);
 
-  const post = new Post({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user._id,
-  });
+  if (!post.url || !post.title) {
+    return response.status(400).send({ error: 'title or url missing ' });
+  }
+
+  if (!post.likes) {
+    post.likes = 0;
+  }
+
+  post.user = user;
 
   const savedPost = await post.save();
+
   user.posts = user.posts.concat(savedPost._id);
   await user.save();
 
-  response.json(savedPost.toJSON());
+  response.json(savedPost);
 });
 
 postsRouter.delete('/:id', async (request, response) => {
@@ -48,8 +43,13 @@ postsRouter.delete('/:id', async (request, response) => {
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' });
   }
+
   const post = await Post.findById(request.params.id);
   const user = await User.findById(decodedToken.id);
+
+  if (!post) {
+    return response.status(404).json({ error: 'blog post not found' });
+  }
 
   if (post.user.toString() === user.id.toString()) {
     await Post.findByIdAndRemove(request.params.id);
@@ -62,95 +62,21 @@ postsRouter.delete('/:id', async (request, response) => {
 });
 
 postsRouter.put('/:id', async (request, response) => {
-  const body = request.body;
-
-  const post = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-  };
+  const post = request.body;
 
   const updatedPost = await Post.findByIdAndUpdate(request.params.id, post, {
     new: true,
-  });
-  response.json(updatedPost.toJSON());
+  }).populate('user', { username: 1, name: 1 });
+  response.json(updatedPost);
 });
 
-module.exports = postsRouter;
- */
-
-// MALLIVASTAUS
-
-const postsRouter = require('express').Router();
-const jwt = require('jsonwebtoken');
-const Blog = require('../models/blog');
-const User = require('../models/user');
-
-postsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
-
-  response.json(blogs);
-});
-
-postsRouter.delete('/:id', async (request, response) => {
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' });
+postsRouter.get('/:id', async (request, response) => {
+  const post = await Post.findById(request.params.id);
+  if (post) {
+    response.json(post);
+  } else {
+    response.status(404).end();
   }
-
-  const user = await User.findById(decodedToken.id);
-  const blog = await Blog.findById(request.params.id);
-  if (blog.user.toString() !== user.id.toString()) {
-    return response
-      .status(401)
-      .json({ error: 'only the creator can delete blogs' });
-  }
-
-  await blog.remove();
-  user.blogs = user.blogs.filter(
-    (b) => b.id.toString() !== request.params.id.toString()
-  );
-  await user.save();
-  response.status(204).end();
-});
-
-postsRouter.put('/:id', async (request, response) => {
-  const blog = request.body;
-
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
-    new: true,
-  });
-  response.json(updatedBlog.toJSON());
-});
-
-postsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body);
-
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' });
-  }
-
-  const user = await User.findById(decodedToken.id);
-
-  if (!blog.url || !blog.title) {
-    return response.status(400).send({ error: 'title or url missing ' });
-  }
-
-  if (!blog.likes) {
-    blog.likes = 0;
-  }
-
-  blog.user = user;
-  const savedBlog = await blog.save();
-
-  user.blogs = user.blogs.concat(savedBlog._id);
-  await user.save();
-
-  response.status(201).json(savedBlog);
 });
 
 module.exports = postsRouter;
